refactor: Menu 테스트 코드 정리, MenuValidator 클래스를 service 패키지로 이동
